:art: Implementa copia de certificados no startup da aplicação
