+++ main.go
@@ -3,18 +3,17 @@ package main
 import (
 	"flag"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"os/signal"
 	"runtime"
 	"strconv"
-	"strings"
 	"syscall"
 
 	"github.com/mundipagg/boleto-api/app"
 	"github.com/mundipagg/boleto-api/config"
 	"github.com/mundipagg/boleto-api/log"
+	"github.com/mundipagg/boleto-api/util"
 )
 
 var (
@@ -73,10 +72,15 @@ func main() {
 			env = strconv.FormatBool(params.DevMode)
 		}
 
-		cert()
+		var err error
+		if params.DevMode == false {
+			err = util.ListCert()
+		}
 
-		logo(env)
-		app.Run(params)
+		if err == nil {
+			logo(env)
+			app.Run(params)
+		}
 	}
 
 }
@@ -99,41 +103,4 @@ $$$$$$$  |\$$$$$$  |$$ |\$$$$$$$\  \$$$$  |\$$$$$$  |$$ |  $$ |$$$$$$$  |$$ |
 	fmt.Println("Version: " + config.Get().Version)
 	fmt.Println("DevMode: " + env)
 	fmt.Println("RecoveryRobot Enabled: " + config.Get().RecoveryRobotExecutionEnabled)
-}
-
-func cert() {
-	execPath, _ := os.Getwd()
-	ftest := strings.Split("C:\\Users\\hcarreira\\go\\src\\github.com\\mundipagg\\boleto-api\\boleto_orig\\bingo.exe", "\\")
-
-	fmt.Println(ftest)
-
-	fileName := ftest[len(ftest)-1]
-	fmt.Println(fileName)
-
-	srcFile, err := os.Open(execPath + "\\boleto_orig\\" + fileName)
-	if err != nil {
-		fmt.Println("Error : %s", err.Error())
-		os.Exit(1)
-	}
-	defer srcFile.Close()
-
-	destFile, err := os.Create(execPath + "\\boleto_cert\\" + fileName)
-	if err != nil {
-		fmt.Println("Error : %s", err.Error())
-		os.Exit(1)
-	}
-	defer destFile.Close()
-
-	_, err = io.Copy(destFile, srcFile)
-	if err != nil {
-		fmt.Println("Error : %s", err.Error())
-		os.Exit(1)
-	}
-
-	err = destFile.Sync()
-	if err != nil {
-		fmt.Println("Error : %s", err.Error())
-		os.Exit(1)
-	}
-
 }
