+++ util/http.go
@@ -1,9 +1,13 @@
 package util
 
 import (
+	"crypto"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/base64"
+	"encoding/pem"
+	"errors"
+	"fmt"
 	"io/ioutil"
 	"net"
 	"net/http"
@@ -11,6 +15,7 @@ import (
 	"time"
 
 	s "github.com/fullsailor/pkcs7"
+	"github.com/mundipagg/boleto-api/config"
 )
 
 var defaultDialer = &net.Dialer{Timeout: 16 * time.Second, KeepAlive: 16 * time.Second}
@@ -121,6 +126,61 @@ func SignRequest(request string) (string, error) {
 	signedRequest := base64.StdEncoding.EncodeToString(detachedSignature)
 
 	return signedRequest, nil
+}
+
+//Read privatekey and parse to PKCS#1
+func parsePrivateKey() (crypto.PrivateKey, error) {
+
+	pkeyBytes, err := ioutil.ReadFile(config.Get().CertICP_PathPkey)
+	if err != nil {
+		return nil, err
+	}
+
+	block, _ := pem.Decode(pkeyBytes)
+	if block == nil {
+		return nil, errors.New("Key Not Found")
+	}
+
+	switch block.Type {
+	case "RSA PRIVATE KEY":
+		rsa, err := x509.ParsePKCS1PrivateKey(block.Bytes)
+		if err != nil {
+			return nil, err
+		}
+		return rsa, nil
+	default:
+		return nil, fmt.Errorf("SSH: Unsupported key type %q", block.Type)
+	}
+
+}
+
+///Read chainCertificates and adapter to x509.Certificate
+func parseChainCertificates() (*x509.Certificate, error) {
+
+	chainCertsBytes, err := ioutil.ReadFile(config.Get().CertICP_PathChainCertificates)
+	if err != nil {
+		return nil, err
+	}
+
+	block, _ := pem.Decode(chainCertsBytes)
+	if block == nil {
+		return nil, errors.New("Key Not Found")
+	}
+
+	cert, err := x509.ParseCertificate(block.Bytes)
+	if err != nil {
+		return nil, err
+	}
+
+	return cert, nil
+}
+
+// Read signedData and parse to *x509.Certificate
+func parseSignedData(request string) (*s.SignedData, error) {
+
+	sig, err := s.NewSignedData([]byte(request))
+
+	return sig, err
 }
 
 func doRequestTLS(method, url, body, timeout string, header map[string]string, transport *http.Transport) (string, int, error) {
